avoid integer division in tests
These two tests created an array of doubles where elements are
N / K where both N and K are integers, which is a "narrowing"
error under the C++11 standard. K is now changed to a double.

+++ src/test/prob/distributions/univariate/continuous/scaled_inv_chi_square_test.cpp
@@ -11,7 +11,7 @@ TEST(ProbDistributionsScaledInvChiSquare, random) {
 TEST(ProbDistributionsScaledInvChiSquare, chiSquareGoodnessFitTest) {
   boost::random::mt19937 rng;
   int N = 10000;
-  int K = 5;
+  double K = 5;
   boost::math::inverse_chi_squared_distribution<>dist (2.0);
   boost::math::chi_squared mydist(K-1);
